integrate cachefs, mergefs, and genfs into overlay
