--- internal/fscache/entry.go
@@ -1,9 +1,0 @@
-package fscache
-
-import (
-	"io/fs"
-)
-
-type Entry interface {
-	open() fs.File
-}
