+++ package/overlay/dir.go
@@ -3,7 +3,7 @@ package overlay
 import (
 	"context"
 
-	"github.com/livebud/bud/package/conjure"
+	"github.com/livebud/bud/package/budfs/genfs"
 )
 
 type DirGenerator interface {
@@ -12,11 +12,11 @@ type DirGenerator interface {
 
 type Dir struct {
 	fsys F
-	*conjure.Dir
+	*genfs.Dir
 }
 
 func (d *Dir) GenerateFile(path string, fn func(ctx context.Context, fsys F, file *File) error) {
-	d.Dir.GenerateFile(path, func(file *conjure.File) error {
+	d.Dir.GenerateFile(path, func(file *genfs.File) error {
 		return fn(context.TODO(), d.fsys, &File{file})
 	})
 }
@@ -26,7 +26,7 @@ func (d *Dir) FileGenerator(path string, generator FileGenerator) {
 }
 
 func (d *Dir) GenerateDir(path string, fn func(ctx context.Context, fsys F, dir *Dir) error) {
-	d.Dir.GenerateDir(path, func(dir *conjure.Dir) error {
+	d.Dir.GenerateDir(path, func(dir *genfs.Dir) error {
 		return fn(context.TODO(), d.fsys, &Dir{d.fsys, dir})
 	})
 }
