+++ package/overlay/embed.go
@@ -3,18 +3,18 @@ package overlay
 import (
 	"context"
 
-	"github.com/livebud/bud/package/conjure"
+	"github.com/livebud/bud/package/budfs/genfs"
 )
 
-type Embed conjure.Embed
+type Embed genfs.EmbedFile
 
 var _ FileGenerator = (*Embed)(nil)
 var _ FileServer = (*Embed)(nil)
 
 func (e *Embed) GenerateFile(_ context.Context, _ F, file *File) error {
-	return (*conjure.Embed)(e).GenerateFile(file.File)
+	return (*genfs.EmbedFile)(e).GenerateFile(file.File)
 }
 
 func (e *Embed) ServeFile(_ context.Context, _ F, file *File) error {
-	return (*conjure.Embed)(e).ServeFile(file.File)
+	return (*genfs.EmbedFile)(e).GenerateFile(file.File)
 }
