--- package/overlay/subscribe.go
@@ -1,5 +1,0 @@
-package overlay
-
-// func (f *FileSystem) Subscribe(name string) pubsub.Subscription {
-// 	return f.ps.Subscribe(name)
-// }
